backend/pyproject.toml
[project]
name = "uv-fastapi-example"
version = "0.1.0"
description = "Hello World FastAPI app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.112.1",
    "uvicorn[standard]>=0.29.0",
    "langchain-core>=0.3.65",
    "langchain-openai>=0.3.22",
    "langgraph>=0.4.8",
    "langgraph-supervisor>=0.0.27",
    "python-dotenv>=1.1.0",
    "langchain-community>=0.3.0",
    "faiss-cpu>=1.8.0",
    "tiktoken>=0.8.0",
    "pydantic-settings>=2.9.1",
    "pymupdf>=1.26.1",
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.19.0",
    "azure-search-documents>=11.4.0",
    "azure-cosmos>=4.5.0",
    "azure-monitor-opentelemetry>=1.2.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-openai>=0.21.0",
]
